requirements.txt
selenium
plistlib
Pillow
scikit-image
matplotlib
scipy
numpy